# Reuse write request from distributor to Ingesters