--- a/pkg/cortexpb/slicesPool.go
+++ b/pkg/cortexpb/slicesPool.go
@@ -21,25 +21,28 @@ func (sp *byteSlicePools) init(pools int) {
 		size := int(math.Pow(2, float64(i)))
 		sp.pools[i] = sync.Pool{
 			New: func() interface{} {
-				return make([]byte, 0, size)
+				buf := make([]byte, 0, size)
+				return &buf
 			},
 		}
 	}
 }
 
-func (sp *byteSlicePools) getSlice(size int) []byte {
+func (sp *byteSlicePools) getSlice(size int) *[]byte {
 	index := int(math.Ceil(math.Log2(float64(size))))
 
 	if index < 0 || index >= len(sp.pools) {
-		return make([]byte, size)
+		buf := make([]byte, size)
+		return &buf
 	}
 
-	s := sp.pools[index].Get().([]byte)
-	return s[:size]
+	s := sp.pools[index].Get().(*[]byte)
+	*s = (*s)[:size]
+	return s
 }
 
-func (sp *byteSlicePools) reuseSlice(s []byte) {
-	index := int(math.Floor(math.Log2(float64(cap(s)))))
+func (sp *byteSlicePools) reuseSlice(s *[]byte) {
+	index := int(math.Floor(math.Log2(float64(cap(*s)))))
 
 	if index < 0 || index >= len(sp.pools) {
 		return