--- a/pkg/cortexpb/slicesPool.go
+++ b/pkg/cortexpb/slicesPool.go
@@ -5,6 +5,10 @@ import (
 	"sync"
 )
 
+const (
+	minPoolSizePower = 5
+)
+
 type byteSlicePools struct {
 	pools []sync.Pool
 }
@@ -18,7 +22,7 @@ func newSlicePool(pools int) *byteSlicePools {
 func (sp *byteSlicePools) init(pools int) {
 	sp.pools = make([]sync.Pool, pools)
 	for i := 0; i < pools; i++ {
-		size := int(math.Pow(2, float64(i)))
+		size := int(math.Pow(2, float64(i+minPoolSizePower)))
 		sp.pools[i] = sync.Pool{
 			New: func() interface{} {
 				buf := make([]byte, 0, size)
@@ -29,11 +33,15 @@ func (sp *byteSlicePools) init(pools int) {
 }
 
 func (sp *byteSlicePools) getSlice(size int) *[]byte {
-	index := int(math.Ceil(math.Log2(float64(size))))
+	index := int(math.Ceil(math.Log2(float64(size)))) - minPoolSizePower
 
-	if index < 0 || index >= len(sp.pools) {
+	if index >= len(sp.pools) {
 		buf := make([]byte, size)
 		return &buf
+	}
+
+	if index < 0 {
+		index = 0
 	}
 
 	s := sp.pools[index].Get().(*[]byte)
@@ -42,10 +50,14 @@ func (sp *byteSlicePools) getSlice(size int) *[]byte {
 }
 
 func (sp *byteSlicePools) reuseSlice(s *[]byte) {
-	index := int(math.Floor(math.Log2(float64(cap(*s)))))
+	index := int(math.Floor(math.Log2(float64(cap(*s))))) - minPoolSizePower
 
-	if index < 0 || index >= len(sp.pools) {
+	if index >= len(sp.pools) {
 		return
+	}
+
+	if index < 0 {
+		index = 0
 	}
 
 	sp.pools[index].Put(s)