--- a/pkg/cortexpb/slicesPool_test.go
+++ b/pkg/cortexpb/slicesPool_test.go
@@ -1,0 +1,18 @@
+package cortexpb
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestByteSlicePools(t *testing.T) {
+	sut := NewSlicePool(20)
+
+	for i := 0; i < 1024*1024; i = i + 128 {
+		s := sut.getSlice(i)
+		assert.Equal(t, len(s), i)
+		sut.reuseSlice(s)
+	}
+
+}