--- a/pkg/cortexpb/timeseries_test.go
+++ b/pkg/cortexpb/timeseries_test.go
@@ -1,8 +1,10 @@
 package cortexpb
 
 import (
+	"fmt"
 	"testing"
 
+	"github.com/gogo/protobuf/proto"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -63,4 +65,57 @@ func TestTimeseriesFromPool(t *testing.T) {
 		assert.Len(t, reused.Labels, 0)
 		assert.Len(t, reused.Samples, 0)
 	})
+}
+
+func BenchmarkMarshallWriteRequest(b *testing.B) {
+	ts := PreallocTimeseriesSliceFromPool()
+
+	for i := 0; i < 100; i++ {
+		ts = append(ts, PreallocTimeseries{TimeSeries: TimeseriesFromPool()})
+		ts[i].Labels = []LabelAdapter{
+			{Name: "foo", Value: "bar"},
+			{Name: "very long label name", Value: "very long label value"},
+			{Name: "very long label name 2", Value: "very long label value 2"},
+			{Name: "very long label name 3", Value: "very long label value 3"},
+			{Name: "int", Value: fmt.Sprint(i)},
+		}
+		ts[i].Samples = []Sample{{Value: 1, TimestampMs: 2}}
+	}
+
+	tests := []struct {
+		name                string
+		writeRequestFactory func() proto.Marshaler
+		clean               func(in interface{})
+	}{
+		{
+			name: "no-pool",
+			writeRequestFactory: func() proto.Marshaler {
+				return &WriteRequest{Timeseries: ts}
+			},
+			clean: func(in interface{}) {},
+		},
+		{
+			name: "pool",
+			writeRequestFactory: func() proto.Marshaler {
+				w := PreallocWriteRequestFromPool()
+				w.Timeseries = ts
+				return w
+			},
+			clean: func(in interface{}) {
+				ReuseWriteRequest(in.(*PreallocWriteRequest))
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		b.Run(tc.name, func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				w := tc.writeRequestFactory()
+				_, err := w.Marshal()
+				require.NoError(b, err)
+				tc.clean(w)
+			}
+			b.ReportAllocs()
+		})
+	}
 }