@alanprot 2023-03-06
> Can you add a Benchmark test for this?

Yup! will do