@alanprot 2023-03-07
Here is the branchmark: We are basically avoiding reallocating the byte[] when serializing.

```
BenchmarkMarshallWriteRequest/no-pool-12         	   51582	     23269 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/no-pool-12         	   53248	     22926 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/no-pool-12         	   52886	     22982 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/no-pool-12         	   51668	     23563 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/no-pool-12         	   51082	     22742 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   57105	     20625 ns/op	      24 B/op	       1 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   58096	     20151 ns/op	      24 B/op	       1 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   59668	     20219 ns/op	      24 B/op	       1 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   59752	     20385 ns/op	      24 B/op	       1 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   59164	     20557 ns/op	      24 B/op	       1 allocs/op
```