@alanprot 2023-03-07
I could make no have any allocation at all:

```
BenchmarkMarshallWriteRequest/no-pool-12         	   50653	     23301 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/no-pool-12         	   46936	     24360 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/no-pool-12         	   52663	     24185 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/no-pool-12         	   52234	     24141 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/no-pool-12         	   51400	     23053 ns/op	   19136 B/op	       2 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   58034	     21088 ns/op	       0 B/op	       0 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   58436	     20775 ns/op	       0 B/op	       0 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   58020	     20741 ns/op	       0 B/op	       0 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   57319	     20861 ns/op	       0 B/op	       0 allocs/op
BenchmarkMarshallWriteRequest/pool-12            	   54622	     23412 ns/op	       0 B/op	       0 allocs/op
```